move id migrate to init
