+++ tests/components/comelit/snapshots/test_diagnostics.ambr
@@ -115,7 +115,7 @@
       'discovery_keys': dict({
       }),
       'domain': 'comelit',
-      'minor_version': 1,
+      'minor_version': 2,
       'options': dict({
       }),
       'pref_disable_new_entities': False,
@@ -181,7 +181,7 @@
       'discovery_keys': dict({
       }),
       'domain': 'comelit',
-      'minor_version': 1,
+      'minor_version': 2,
       'options': dict({
       }),
       'pref_disable_new_entities': False,
