+++ tests/components/comelit/test_sensor.py
@@ -16,13 +16,12 @@ from aiocomelit.const import (
     AlarmZoneState,
 )
 from freezegun.api import FrozenDateTimeFactory
-import pytest
 from syrupy.assertion import SnapshotAssertion
 
-from homeassistant.components.comelit.const import DOMAIN, SCAN_INTERVAL
+from homeassistant.components.comelit.const import SCAN_INTERVAL
 from homeassistant.const import STATE_UNKNOWN, Platform
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers import device_registry as dr, entity_registry as er
+from homeassistant.helpers import entity_registry as er
 
 from . import setup_integration
 
@@ -190,108 +189,3 @@ async def test_vedo_sensor_dynamic(
 
     assert hass.states.get(ENTITY_ID)
     assert hass.states.get(entity_id_2)
-
-
-@pytest.mark.parametrize(
-    (
-        "device_identifier_type",
-        "old_unique_id_suffix",
-        "expected_unique_id_suffix",
-        "old_unique_id_removed",
-    ),
-    [
-        (
-            "zone",
-            "0",
-            "human_status-0",
-            True,
-        ),
-        (
-            "other",
-            "0",
-            "0",
-            False,
-        ),
-        (
-            "zone",
-            "human_status-0",
-            "human_status-0",
-            False,
-        ),
-    ],
-)
-async def test_migrate_sensor_unique_id(
-    hass: HomeAssistant,
-    mock_vedo: AsyncMock,
-    mock_vedo_config_entry: MockConfigEntry,
-    device_identifier_type: str,
-    old_unique_id_suffix: str,
-    expected_unique_id_suffix: str,
-    old_unique_id_removed: bool,
-    device_registry: dr.DeviceRegistry,
-    entity_registry: er.EntityRegistry,
-) -> None:
-    """Test sensor unique ID migration."""
-    config_entry = mock_vedo_config_entry
-    config_entry.add_to_hass(hass)
-
-    old_unique_id = f"{config_entry.entry_id}-{old_unique_id_suffix}"
-    new_unique_id = f"{config_entry.entry_id}-{expected_unique_id_suffix}"
-
-    device = device_registry.async_get_or_create(
-        config_entry_id=config_entry.entry_id,
-        identifiers={(DOMAIN, f"{config_entry.entry_id}-{device_identifier_type}-0")},
-    )
-
-    entity_entry = entity_registry.async_get_or_create(
-        Platform.SENSOR,
-        DOMAIN,
-        old_unique_id,
-        config_entry=config_entry,
-        device_id=device.id,
-    )
-
-    await setup_integration(hass, config_entry)
-
-    migrated_entry = entity_registry.async_get(entity_entry.entity_id)
-    assert migrated_entry
-    assert migrated_entry.unique_id == new_unique_id
-    old_entity_id = entity_registry.async_get_entity_id(
-        Platform.SENSOR, DOMAIN, old_unique_id
-    )
-    assert (old_entity_id is None) is old_unique_id_removed
-    assert (old_entity_id == entity_entry.entity_id) is not old_unique_id_removed
-
-
-async def test_migrate_sensor_unique_id_missing_device(
-    hass: HomeAssistant,
-    mock_vedo: AsyncMock,
-    mock_vedo_config_entry: MockConfigEntry,
-    device_registry: dr.DeviceRegistry,
-    entity_registry: er.EntityRegistry,
-) -> None:
-    """Test sensor unique ID migration when linked device no longer exists."""
-    config_entry = mock_vedo_config_entry
-    config_entry.add_to_hass(hass)
-
-    old_unique_id = f"{config_entry.entry_id}-0"
-
-    device = device_registry.async_get_or_create(
-        config_entry_id=config_entry.entry_id,
-        identifiers={(DOMAIN, f"{config_entry.entry_id}-zone-0")},
-    )
-
-    entity_entry = entity_registry.async_get_or_create(
-        Platform.SENSOR,
-        DOMAIN,
-        old_unique_id,
-        config_entry=config_entry,
-        device_id=device.id,
-    )
-
-    with patch.object(device_registry, "async_get", return_value=None):
-        await setup_integration(hass, config_entry)
-
-    migrated_entry = entity_registry.async_get(entity_entry.entity_id)
-    assert migrated_entry
-    assert migrated_entry.unique_id == old_unique_id
